Use center-symmetric census transform

+++ src/types.hpp
@@ -21,7 +21,7 @@ limitations under the License.
 
 namespace sgm {
 
-using feature_type = uint64_t;
+using feature_type = uint32_t;
 using cost_type = uint8_t;
 using cost_sum_type = uint16_t;
 using output_type = uint8_t;
